% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hr_employee_data.R
\docType{data}
\name{hr_employee_data}
\alias{hr_employee_data}
\title{HR employee data}
\format{
A dataframe with 814 rows and 8 variables:
\describe{
\item{Gender}{The gender of the employee}
\item{Department}{The department the employee works in}
\item{Annual Salary}{The employee's annual salary}
\item{Location}{The city where the company office is located}
\item{Rating}{The employee's last performance rating (1 - Very Poor; 2 - Poor; 3 - Average; 4 - Good; 5 - Very Good)}
\item{Distance to Office}{The distance from the employee's home to the office (in kilometers)}
\item{Age}{The employee's age}
\item{Tenure in Company}{The number of years working at the company}
}
}
\usage{
hr_employee_data
}
\description{
Fictional hr employee dataset
}
\examples{
hr_employee_data
}
\keyword{datasets}
